% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BACON_analysis.R
\docType{class}
\name{BACON-class}
\alias{BACON-class}
\alias{BACON}
\title{The key slots used in the BACON object are described below}
\description{
The key slots used in the BACON object are described below
}
\section{Slots}{

\describe{
\item{\code{data.raw}}{raw count data matrix}

\item{\code{data}}{gene count matrix for BACON analysis (Genes should be in rows and bacteria in columns)}

\item{\code{data_qs}}{a data frame only containing QS related genes and bacteria sub-populations or species labels}

\item{\code{net0}}{a list containing the original communication strength matrix without filtered by p-value(0.05)}

\item{\code{pvalue}}{a list containing the p-value (calculated by permutation test) matrix corresponding to net0}

\item{\code{net}}{a list of containing the communication strength matrix (number of synthesis bacteria groups X number of reception bacteria groups) filted by p-value. Each row of the communication strength matrix indicates the communication probability originating from the synthesis bacteria group to other groups.}

\item{\code{fc}}{deprecated}

\item{\code{info}}{information of each synthesis-reception pair}

\item{\code{ligand.abundance}}{ligand abundance of synthesis bacteria groups(row) for synthesis-reception pair (column)}

\item{\code{target.abundance}}{target abundance of reception bacteria groups(row) for synthesis-reception pair (column)}

\item{\code{meta}}{data frame containing information associated with each bacteria}

\item{\code{idents}}{a factor defining the cell identity used for all analysis. It becomes a list for a merged BACON object}

\item{\code{DB}}{synthesis-reception communication database used in the analysis}

\item{\code{LR}}{a list of information related with synthesis-reception communication pairs}
}}

