% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaconClass.R, R/BaconMethods.R
\docType{methods}
\name{bias}
\alias{bias}
\alias{bias,Bacon-method}
\alias{bias}
\title{Method to extract the estimated bias from the 'bacon'-object}
\usage{
bias(object)

\S4method{bias}{Bacon}(object)
}
\arguments{
\item{object}{'bacon'-object}
}
\value{
vector or matrix of inflation
}
\description{
Method to extract the estimated bias from the 'bacon'-object
}
\examples{
y <- rnormmix(2000, c(0.9, 0, 1, 0, 4, 1))
##nbins = 100 to speed up the calculations
bc <- bacon(y, nbins=100)
bias(bc)
}
\seealso{
\code{\link{bacon}}
}
