% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BACON_analysis.R
\name{createBACON}
\alias{createBACON}
\title{Create a new BACON object from a data matrix}
\usage{
createBACON(object, meta = NULL, group.by = NULL, database = NULL)
}
\arguments{
\item{object}{a gene count data matrix (rownames = gene names; colnames = bacteria)}

\item{meta}{a data frame (rownames = bacteria barcode) consisting of bacteria information, which will be used for defining baceria groups}

\item{group.by}{a vector to indicate group annotation of bacteria}

\item{database}{BACON database selected by users}
}
\description{
Create a new BACON object from a data matrix
}
