% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normmixture.R
\name{dnormmix}
\alias{dnormmix}
\title{density of a k-component normal mixture}
\usage{
dnormmix(x, theta)
}
\arguments{
\item{x}{x like dnorm(x, ...}

\item{theta}{parameters of the mixture proportion, mean and sd}
}
\value{
density of a k-component normal mixture
}
\description{
density of a k-component normal mixture
}
\details{
details follow
}
\examples{
n <- 2000
theta <- c(0.8, 0, 1, 0, 4, 1)
x <- rnormmix(n, theta)
hist(x, freq=FALSE, n=100)
curve(dnormmix(x, theta), add=TRUE, lwd=2)
}
\author{
mvaniterson
}
