% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BACON_analysis.R
\name{expression_calculate_by_select_group}
\alias{expression_calculate_by_select_group}
\title{Calculate the average gene expression by bacterial sub-populations or species}
\usage{
expression_calculate_by_select_group(qs_df, gene_used, calculate_method = NULL)
}
\arguments{
\item{qs_df}{data frame containing QS-related gene expression (rownames = bacteria; colnames = genes;the last column is bacteria group labels)}

\item{gene_used}{gene symbols used to calculate average expression for bacteria groups}

\item{calculate_method}{set "mean" as default}
}
\description{
Calculate the average gene expression by bacterial sub-populations or species
}
