% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BACON_analysis.R
\name{get_orig_strength_matrix}
\alias{get_orig_strength_matrix}
\title{Calculate the communication strength matrix for a single synthesis-reception pair (without permutation test)}
\usage{
get_orig_strength_matrix(
  qs_df,
  ligand_cell,
  target_cell,
  synthsis_complex_new,
  ligand_complex,
  ligand_complex_number,
  receptor_complex_new,
  target_complex,
  target_complex_number,
  calculate_method = NULL
)
}
\arguments{
\item{qs_df}{data frame containing QS-related gene expression (rownames = bacteria; colnames = genes;the last column is bacteria group labels)}

\item{ligand_cell}{synthesis bacteria groups}

\item{target_cell}{reception bacteria groups}

\item{synthsis_complex_new}{vector with updated gene symbols (removing those excluded from the expression data) that are regarded as synthesis complex}

\item{ligand_complex}{vector indicating the groups of synthesis complex}

\item{ligand_complex_number}{vector indicating the stoichiometry of each group of synthesis complex for calculating ligand abundance}

\item{receptor_complex_new}{vector with updated gene symbols (removing those excluded from the expression data) that are regarded as receptor complex}

\item{target_complex}{vector indicating the groups of receptor complex}

\item{target_complex_number}{vector indicating the stoichiometry of each group of receptor complex for calculating target abundance}
}
\description{
Calculate the communication strength matrix for a single synthesis-reception pair (without permutation test)
}
