% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaconClass.R, R/BaconMethods.R
\docType{methods}
\name{meta}
\alias{meta}
\alias{meta,Bacon-method}
\title{fixed meta-analysis}
\usage{
meta(object, corrected = TRUE, ...)

\S4method{meta}{Bacon}(object, corrected = TRUE, ...)
}
\arguments{
\item{object}{'bacon'-object}

\item{corrected}{optional return uncorrected}

\item{...}{additional arguments}
}
\value{
list of fixed-effect meta analysis output
}
\description{
Perform fixed meta-analysis using inflation and bias corrected
effect-sizes and standard errors
}
\details{
TODO maybe add idea's from http://www.netstorm.be/home/meta_analysis#metaAnalysisU
}
\seealso{
\code{\link{bacon}}
}

