% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BACON_analysis.R
\name{net_aggregation}
\alias{net_aggregation}
\title{Aggregation of communication networks over all synthesis-reception pairs}
\usage{
net_aggregation(net_list)
}
\arguments{
\item{net_list}{list containing communication strength matrix for all synthesis-reception pairs}
}
\description{
Aggregation of communication networks over all synthesis-reception pairs
}
