% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BACON_visual.R
\name{netVisual_chord_BACON}
\alias{netVisual_chord_BACON}
\title{Chord diagram of QS communication for single or multiple communication pairs}
\usage{
netVisual_chord_BACON(
  object,
  interaction_use = "all",
  synthesis_subpopulation = NULL,
  reception_subpopulation = NULL,
  lab.cex = 1.3
)
}
\arguments{
\item{object}{a BACON object}

\item{interaction_use}{QS communication indexes used for aggregation ("all" means use all communication pairs); for a single communication pair,set interaction_use as the communication name}

\item{synthesis_subpopulation}{synthesis group used for plot}

\item{reception_subpopulation}{reception group used for plot}
}
\description{
Chord diagram of QS communication for single or multiple communication pairs
}
