% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BACON_visual.R
\name{netVisual_chord_cell_internal}
\alias{netVisual_chord_cell_internal}
\title{Chord diagram for visualizing QS communication from a weighted adjacency matrix}
\usage{
netVisual_chord_cell_internal(
  net,
  color.use = NULL,
  group = NULL,
  cell.order = NULL,
  sources.use = NULL,
  targets.use = NULL,
  lab.cex = 0.8,
  small.gap = 1,
  big.gap = 10,
  annotationTrackHeight = c(0.03),
  remove.isolate = FALSE,
  link.visible = TRUE,
  scale = FALSE,
  directional = 1,
  link.target.prop = TRUE,
  reduce = -1,
  transparency = 0.4,
  link.border = NA,
  title.name = NULL,
  show.legend = FALSE,
  legend.pos.x = 20,
  legend.pos.y = 20,
  ...
)
}
\arguments{
\item{net}{weighted matrix with columns defining the QS network}

\item{color.use}{color use for groups}

\item{group}{group labels the sector name should be used as the names in the vector}

\item{cell.order}{defining the bacteria type orders}

\item{sources.use}{vector giving the index or the name of source bacterial groups}

\item{targets.use}{vector giving the index or the name of target bacterial groups}

\item{lab.cex}{font size for the text}

\item{small.gap}{small gap between sectors}

\item{big.gap}{Gap between the different sets of sectors, which are defined in the \code{group} parameter}

\item{annotationTrackHeight}{annotationTrack height}

\item{remove.isolate}{whether remove sectors without any links}

\item{link.visible}{whether plot the link. The value is logical, if it is set to FALSE, the corresponding link will not plotted, but the space is still ocuppied.}

\item{scale}{scale each sector to same width}

\item{directional}{Whether links have directions. 1 means the direction is from the first column in df to the second column, -1 is the reverse, 0 is no direction, and 2 for two directional.}

\item{link.target.prop}{If the Chord diagram is directional, for each source sector, whether to draw bars that shows the proportion of target sectors}

\item{reduce}{if the ratio of the width of certain grid compared to the whole circle is less than this value, the grid is removed on the plot. Set it to value less than zero if you want to keep all tiny grid.}

\item{transparency}{transparency of link colors}

\item{link.border}{border for links, single vector or a matrix with names or a data frame with source, target , prob, three column information}

\item{title.name}{title name of the plot}

\item{show.legend}{whether show the figure legend}

\item{legend.pos.x, legend.pos.y}{adjust the legend position}
}
\description{
Chord diagram for visualizing QS communication from a weighted adjacency matrix
}
