% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BACON_visual.R
\name{netVisual_circle_BACON}
\alias{netVisual_circle_BACON}
\title{Circle plot for QS communication network}
\usage{
netVisual_circle_BACON(
  net_ori,
  color.use = NULL,
  title.name = NULL,
  synthesis_sub = NULL,
  reception_sub = NULL,
  remove.isolate = FALSE,
  top = 1,
  weight.scale = TRUE,
  vertex.weight = 1,
  vertex.weight.max = NULL,
  vertex.size.max = NULL,
  vertex.label.cex = 1.2,
  vertex.label.color = "black",
  edge.weight.max = NULL,
  edge.width.max = 5,
  alpha.edge = 0.6,
  label.edge = FALSE,
  edge.label.color = "black",
  edge.label.cex = 0.8,
  edge.curved = 0.2,
  shape = NULL,
  layout = in_circle(),
  margin = 0.2,
  vertex.size = NULL,
  arrow.width = 1,
  arrow.size = 0.8
)
}
\arguments{
\item{net_ori}{A weighted matrix representing the connections}

\item{title.name}{the name of the title}

\item{synthesis_sub}{vector giving the index or the name of the ligand bacterial groups}

\item{reception_sub}{vector giving the index or the name of the target bacterial groups}

\item{remove.isolate}{whether remove the isolate nodes in the communication network}

\item{top}{the fraction of communications to show}

\item{weight.scale}{whether scale the weight}

\item{vertex.weight}{the weight of vertex:either a scale value or a vector}

\item{vertex.weight.max}{the maximum weight of vertex; defualt = max(vertex.weight)}

\item{vertex.size.max}{the maximum vertex size for visualization}

\item{vertex.label.cex}{The label size of vertex}

\item{vertex.label.color}{The color of label for vertex}

\item{edge.weight.max}{the maximum weight of edge; defualt = max(net)}

\item{edge.width.max}{The maximum edge width for visualization}

\item{alpha.edge}{the transprency of edge}

\item{label.edge}{Whether or not shows the label of edges}

\item{edge.label.color}{The color for single arrow}

\item{edge.label.cex}{The size of label for arrows}

\item{edge.curved}{Specifies whether to draw curved edges, or not.}

\item{arrow.width}{The width of arrows}

\item{arrow.size}{the size of arrow}
}
\description{
Circle plot for QS communication network
}
