% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normmixture.R
\name{plotnormmix}
\alias{plotnormmix}
\title{plot normal mixtures}
\usage{
plotnormmix(x, theta, ...)
}
\arguments{
\item{x}{vector of test statistics}

\item{theta}{parameters describing the mixture components}

\item{...}{arguments passed to hist}
}
\value{
return plot with histogram of the data and mixture and individual components
}
\description{
plot normal mixtures
}
\details{
details follow
}
\examples{
n <- 2000
theta <- c(0.8, 0, 1, 0, 4, 1)
x <- rnormmix(n, theta)
plotnormmix(x, theta)
}
\author{
mvaniterson
}
