% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normmixture.R
\name{rnormmix}
\alias{rnormmix}
\title{sample from a normal mixture}
\usage{
rnormmix(n, theta)
}
\arguments{
\item{n}{size}

\item{theta}{parameters}
}
\value{
n samples from a normal mixture with parameters theta
}
\description{
sample from a normal mixture
}
\details{
details follow
}
\examples{
n <- 2000
theta <- c(0.8, 0, 1, 0, 4, 1)
x <- rnormmix(n, theta)
}
\author{
mvaniterson
}

