% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BACON_analysis.R
\name{run_BACON}
\alias{run_BACON}
\title{Calculate the communication strength matrix for all synthesis-reception pair}
\usage{
run_BACON(
  object = object,
  ligand_cell = NULL,
  target_cell = NULL,
  M = 100,
  calculate_method = NULL
)
}
\arguments{
\item{object}{a BACON object}

\item{ligand_cell}{synthesis bacteria groups}

\item{target_cell}{reception bacteria groups}

\item{M}{number of permutation tests}
}
\description{
Calculate the communication strength matrix for all synthesis-reception pair
}
