% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BACON_visual.R
\name{scPalette}
\alias{scPalette}
\title{Generate color from a customed color palette}
\usage{
scPalette(n)
}
\arguments{
\item{n}{number of color}
}
\description{
Generate color from a customed color palette
}
