% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaconClass.R
\docType{class}
\name{Bacon-class}
\alias{Bacon-class}
\title{An S4 class container for storing Gibbs Sampler input and output}
\description{
An S4 class container for storing Gibbs Sampler input and output
}
\section{Slots}{

\describe{
\item{\code{teststatistics}}{numeric vector or matrix of test-statistics}

\item{\code{effectsizes}}{numeric vector or matrix of effect-sizes}

\item{\code{standarderrors}}{numeric vector or matrix of standard errors}

\item{\code{traces}}{array of Gibbs Sampler traces}

\item{\code{estimates}}{vector or matrix of parameter estimates}

\item{\code{priors}}{list of parameters of for the prior distributions}

\item{\code{niter}}{number of iterations}

\item{\code{nburnin}}{length of the burnin period}
}}

