% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaconClass.R, R/BaconMethods.R
\docType{methods}
\name{meta}
\alias{meta}
\alias{meta,Bacon-method}
\title{fixed meta-analysis}
\usage{
meta(object, corrected = TRUE, ...)

\S4method{meta}{Bacon}(object, corrected = TRUE, ...)
}
\arguments{
\item{object}{'bacon'-object}

\item{corrected}{optional return uncorrected}

\item{...}{additional arguments}
}
\value{
object of class 'bacon' with added fixed-effect meta-analysis
test-statistics, effect-sizes and standard-errors
}
\description{
Perform fixed meta-analysis using inflation and bias corrected
effect-sizes and standard errors
}
\details{
TODO maybe add idea's from http://www.netstorm.be/home/meta_analysis#metaAnalysisU
}
\examples{
es <- replicate(6, rnormmix(2000, c(0.9, 0, 1, 0, 4, 1)))
se <- replicate(6, 0.8*sqrt(4/rchisq(2000,df=4)))
bc <- bacon(NULL, es, se)
mbc <- meta(bc)
}
\seealso{
\code{\link{bacon}}
}
