% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaconClass.R
\docType{methods}
\name{plot,Bacon-method}
\alias{plot,Bacon-method}
\title{plot hist or qq}
\usage{
\S4method{plot}{Bacon}(x, y, type = c("hist", "qq"))
}
\arguments{
\item{x}{'bacon'-object}

\item{y}{NULL}

\item{type}{hist or qq}
}
\value{
either qq-plot of P-values or histogram of Test-statistics
}
\description{
simple ggplot2 plotting function for 'bacon'-object
}

